Adding abbreviations to dropdown selects
Measurement units and duty expression codes have abbreviations which are
more commonly used in place of the description or code, so makes sense
to use them.

Also took the time to properly align all dropdown codes with their
descriptions so not too much nor too little spacing was between them.