Change additional code saving on measures
The additional codes should be set regardless if there is a gn code.

Splitting the CodesAnalyzer class into two.

It should also expand the list of codes, so if there is one gn code and 2 additional code, then 2 measures should be created.

If there 2 gn codes and 2 additional code, then 4 measures should be created.

If there are no gn codes and 1 additonal code, then 1 measure should be created.