Add chromedriver-helper to automate setup
With a clean clone, after running bin/setup, the test suite fails
unless you have chromedriver installed (which the README doesn't
list as a dependency).

Instead of adding a manual setup step, this installation can be
automated. The helper also allows the version of Chromedriver to
be specified, which is useful for resolving Chrome/Driver version
mismatches.

https://github.com/flavorjones/chromedriver-helper