Fix Sequel deprecation warning
Suggested remedy to use a joined dataset as a Sequel::Model is adding #from_self

Warning was:
SEQUEL DEPRECATION WARNING: Using a joined dataset as a Sequel::Model dataset is deprecated and will be removed in Sequel 5.  Call from_self on the dataset to wrap it in a subquery.

Release notes:
https://sequel.jeremyevans.net/rdoc/files/doc/release_notes/4_46_0_txt.html