Fix flaky ME32 measure specs
These specs regularly fail on CI:
https://gitlab.bitzesty.com/clients/trade-tariff/trade-tariff-management/-/jobs/2214
https://gitlab.bitzesty.com/clients/trade-tariff/trade-tariff-management/-/jobs/2224
https://gitlab.bitzesty.com/clients/trade-tariff/trade-tariff-management/-/jobs/2333

After extensive debugging (bisect didn't help) it was possible to establish that
something is mutating the addtional code type from non-meursing (applicable to ME32)
to meursing (not applicable to ME32, so the validation doesn't trip as required by the test).

Setting the association directly, rather than the ID, seems to resolve the issue.