Change rollback worker to not use workbasketsearch class
The workbasket search class filters target workbaskets for the XML exports to ones which are in specific states, where for rollback we do not care and all workbaskets should be rolled back.

NB: with by_date_range, an end date needs to be set or else only workbaskets with the operation date of the start date will be removed.