Only set searchable data columns when refreshing searchable data
Previously setting the searchable data for a measure was updating
columns for any other attributes which happened to have changed on
an instance of a measure.

Although this behaviour wasn't the root cause of a bug, it was surprising
that columns unrelated to searchable data were updated and made debugging
of an issue harder.

The (unrelated) bug which lead to an examination of set searchable data
was:
https://trello.com/c/3YeYochw/650-tq-165-dit-xml-generated-by-tap-should-have-3-in-the-update-type-throughout-the-xml-but-update-type-is-1-in-the-xml-updatetype1