Add bootsnap and remove fileutils gem
fileutils is included in Ruby and doesn't need to be loaded via
bundler. Doing so threw warnings with bootsnap:

warning: already initialized constant FileUtils::VERSION

After having removed this from the gemfile, you may need to run
gem uninstall fileutils to remove the warnings because it'll still
potentially be on your gem path.