Fix asset "fonts-ie8.css" is not present in the pipeline
Since updating govuk_template gem in:
https://gitlab.bitzesty.com/clients/trade-tariff/trade-tariff-management/commit/2d880a6b7aa346e44f1c7785d854b0c717de7a9d

The following error is throughout the test output:

DEPRECATION WARNING: The asset "fonts-ie8.css" is not present in the asset pipeline.Falling back to an asset that may be in the public folder.
This behavior is deprecated and will be removed.
To bypass the asset pipeline and preserve this behavior,
use the `skip_pipeline: true` option.
 (called from _app_views_layouts_govuk_template_html_erb___4004431768922234652_47420536064480 at /builds/clients/trade-tariff/trade-tariff-management/app/views/layouts/govuk_template.html.erb:15)

This is because we dont't use govuk_template.html.erb from the gem,
it's been customised in our project (for better... or worse).

Related commit on govuk_template which introduced the font change:
https://github.com/alphagov/govuk_template/commit/083e00576cf8194150d20eb42297db458b2f098b