Remove all linting from circleci
Due to the current state of the codebase, it is likely that every file will require
a lot of work to be able to pass the linting requirements. This will slow development
down and, with a tight deadline, this is not our current priority so this change
will remove the linting steps from our CI process.