Refactor quota creation spec
Abstract capybara behaviour into methods describing the page on which they are performed
in order to both make the test easier to understand at a glance and make the pinpointing of
failing steps easier conceptually.

Remove magic strings and numbers.

Note: measurement_unit uses the `let!` method because it cannot be lazy evaluated because
it needs to exist before the measurement_unit method is invoked in order to populate the
measurement unit dropdown.