Fix quota period type bug
Regardless of the period type we assign to a quota (i.e annual,
quarterly etc) when we go to edit the quota it shows up as
`custom`.

This PR fixes this bug.

I have added an end period date to the quota factory `actual` trait
as without one it quota definition could not be saved and so the
factory does not represent real data without this change.

Note: It looks like the seed data do not have a `workbasket_type_of_quota`
so additional work will need to be done for the existing data.