ON6 bugfix
ON6 conformance error is getting raised when a user does not fill in the quota periods.

The expected behaviour is that it shows an error that tells the user that they missed filling
in a field and does not show conformance errors until the form is correctly completed.

This PR fixes this bug by only doing conformance checks if the quota period section of the
quota form is filled out and raising general validation errors if it is not.