Allow upload xml from edit measures workbaskets
We have an issue whereby the uploading of an edit measure workbasket is not
allowed despite it being approved. This is because it has a slightly different status,
i.e `awaiting_cds_upload_edit` rather than the expected `awaiting_cds_upload_create_new`.

This commit allows workbaskets with either of these statuses to be uploaded.