Fix: Underlying item crosscheck status changes
Prior to this change, when a workbasket was rejected or approved by
cross-checker only the status of the workbasket changed, not the status
of the underlying items.

This change adds `submit_for_approval!` and `reject_cross_check!`
methods which handle the changing of status of both the workbasket
and the objects within the workbasket.