Fix: Underlying item approval status changes
Prior to this change, when a workbasket was rejected or approved by
approver only the status of the workbasket changed, not the status
of the underlying items.

This change adds `confirm_approval!` and `reject_approval!`
methods which handle the changing of status of both the workbasket
and the objects within the workbasket.