Fix: Underlying item approval status changes
Prior to this change, when a workbasket was rejected or approved using
the tester backdoor only the status of the workbasket changed, not the
status of the underlying items.

This change adds `testing_status_backdoor!`
methods which handle the changing of status of both the workbasket
and the objects within the workbasket.

It also changes `allowed_status` to include `cds_error` rather than
`rejected` as rejected is not currently a status in our app.