Fix: Return XML strings in hash
Prior to this change, the download job returned an array of xml strings.
This makes it difficult to identify which file is a mainfile and which is a
metadata file.

This change returns a hash of filename/xml string pairs to make it
easy to identify which file is which.