Fix: Rejection reason not showing up
Prior to this change, if a workbasket is rejected from approval or cross
check it does not show why it was rejected.

This change fixes this issue by stopping multiple events being created
after cross-check or approval. The reason that the rejection reason is not
showing is because we were creating 2 events everytime we reject or approve
a workbasket.

The first event has the rejection reason included but the second
does not and is overriding the first when shown on page. This commit
stops 2 calls to `move_status_to!` for every reject or approve.

It also has a side benefit of cleaning up the event history.

Trello card: https://trello.com/c/yHV2NLL0/819-submitter-cannot-see-the-the-approval-rejected-reason-for-quota