Fix: Tweak info on `Review and Submit` measures page
Prior to this change, we had some unhelpful or unnecessary info on
the review and submit measures page.

This change removes operation date, which we no longer use, and adds
the measure type id to the measure type field in the details table in order to
help users more easily understand which type they selected.