Fix: Tweak info on `Review and Submit` quotas page
Prior to this change, we had some unhelpful or unnecessary info on
the review and submit measures page.

This change removes operation date, which we no longer use, and tweaks
the way we display quota type to match with the way we display measure
type on the measure details table.