Fix: Bug causing multiple measures to be created
Prior to this change, when we edited a 'edit bulk measure' workbasket
it created multiple measures and sent them all in the xml. The expected
behaviour is that 1 new measure is created for every measure that is edited,
i.e the old one is end dated and the new one is created.

This bug meant that every additional edit added an extra measure. i.e if we
withdrew/edited the workbasket 3 times there would be 4 measures.

This change deletes the previous edit everytime we move a bulk edit measure
workbasket into the editing state so that only the original measure, and one edit
ever exists.