Fix: Clicking back after withdraw/edit bug
Prior to this change, if user withdrew a create quota or create measure
workbasket, then clicked the back button, then clicked view to inspect the
workbasket the page would crash.

This change redirects the user to the editing page if the status of
a workbasket is 'editing' even if the user has gone back a page.