Fix: Remove operation date field from create geographical areas
Prior to this change, there was a mandatory `operation date` field in the
create geographical areas form. This is unused due to us manually sending
xml to cds, rather than doing it based on operation date.

This change removes the operation date field and makes it non
mandatory so as not to raise errors.