Fix: Add/Remove memberships in geo area frontend
Prior to this change, memberships are not being displayed
when we go to edit an existing geographical area and there
is no way to add or remove memberships.

This change allows the user to view memberships and view
the memberships that they are adding or removing in the ui.

Note: This is a front end fix only. Users cannot successfully submit
these changes. Commits to follow to add that functionality.