Fix: Groups adding other groups as members
Prior to this change, groups could add other groups as their members.
This should not be the case, only country or regions should be allowed
to be members of a group.

This change adds a check into the initial validation step of create and edit
geo areas workbaskets in order to stop this bug.