Fix: Incorrect conformance rules for geo area
Prior to this change, the code the conformance rules executed did not relate to
the actual conformance rule.

The rule is 'At least one description record is mandatory. The start date of the first
description period must be equal to the start date of the geographical area. Two
descriptions may not have the same start date. The start date of the description
must be less than or equal to the end date of the geographical area.'

This change means we use the start date of the FIRST description period and ensure
it is equal to the start date of the geo area as described. It also ensures that the START
date of the description must be less than or equal to the end date of the area, previously
it checked the validity end date.