Fix: Eager load lib directory
Prior to this change, lib directory was autoloaded, which is not the
recommended practice and appeard to be causing an issue in production.

This change changes autoloading lib to eager loading (i.e. now loads up
the lib directory at startup)

Tree in lib was not loading when browsing the nomenclature tree
https://uktrade.atlassian.net/browse/TARIFFS-205