Fix: Server errors when editing geo area with no memberships
Prior to this change, if a user tried to edit a geo area with no memberships,
and did not add any memberships, a server error would be raised.

This was due to settings_params['geographical_area_memberships'] being
nil and some lines in the settings saver trying to call methods on nil.

This change corrects this issue.