Fix: Nomenclature tree bug for heading group
Prior to this change, if a Heading item (e.g. 710100000) had entries
for bot producline_suffix 10 AND 80 an exception would be thrown.

This change treats headings with a suffix 10 as a 'heading group title'
See https://www.trade-tariff.service.gov.uk/chapters/71
for an example, where suffix 10 is displayed as a heading "
I. NATURAL OR CULTURED PEARLS AND PRECIOUS OR SEMI-PRECIOUS STONES"

https://uktrade.atlassian.net/browse/TARIFFS-330