Fix: Oauth security issue
This change is needed to mitigate against CVE-2015-9284. Installing the gem and
changing the get requests to our sso to post requests.

Our application running `authenticate_user!` before every controller action
so we cannot simply change the get requests made to our oauth provider to post requests
because we are redirecting all requests there if user is not logged in. Instead I have
added a simple view with a log in button to allow us to make post requests to our
oauth provider.