Fix: Create user from SSO bug
Prior to this change, when a user was created in our db by signing into the dit sso for
the first time the user did not get assigned an id and it caused the application to crash.

This commit assigns an id when creating a new user and fixes the issue.