Fix: Broken page when reviewing/cross-checking/approving quota
Prior to this change, the page would error and crash when we tried to review/approve
or view a submitted quota. This was due to a change to the way we displayed dates.
The summary_of_configuration page was passing a string to the format_date method
instead of passing a date to it.

This commit fixes this issue.