Fix: Nomenclature date setting and search conflict
Prior to this change, the nomenclature date change was using a cookie to
remember the date chosen. This conflicts with the way the original
system used dates (within time_machine) to restrict searches etc.

This change uses a similar mechanism to continue to pass the date
parameters within the url.

https://uktrade.atlassian.net/browse/TARIFFS-378
https://uktrade.atlassian.net/browse/TARIFFS-333