Fix: Unnecessary commodity code warning for create footnotes
Prior to this change, when a user was creating a footnote and selected a footnote
type that could be associated with a goods commodity code, entered an incorrect
commodity code, but then changed to a footnote type that did not need a commodity code
the incorrect commodity code error still showed.

This change fixes the issue by only checking for commodity code errors if the footnote
type is one that takes a commodity code.