Fix: Unnecessary measure warning for create footnotes
Prior to this change, when a user was creating a footnote and selected a footnote
type that could be associated with a measure, entered an incorrect
measure, but then changed to a footnote type that did not need a measure
the incorrect measure error still showed.

This change fixes the issue by only checking for measure errors if the footnote
type is one that takes a measure.

https://uktrade.atlassian.net/jira/software/projects/TARIFFS/boards/127?selectedIssue=TARIFFS-395