Fix: Remove end date option for edit footnotes
Prior to this change, there was a field in which a user could edit the end date of a footnote.

As per Matt's request on the create footnote card we have disabled the ability for a user to
set an end date on a footnote so seems sensible to disable their ability to set one in the edit
mode.

See: https://uktrade.atlassian.net/jira/software/projects/TARIFFS/boards/127?selectedIssue=TARIFFS-130