Refactor: Adding of new footnote description and description period
Prior to this commit, the edit footnote settings saver handled the editing of a footnote
description or description period differently depending on whether the date was in the
future or not. This is unnecessary.

This commit removes this duplication in order to make the code easier to read and
understand.