New: Allow generating of delete XML for footnotes associated measures
This commit allows the generating of xml to delete a footnote measure association.

The changes are complicated by the fact that when updating an object with the operation: 'D' the
oplog records it as an update so the operation changes to 'U'. In order to record a 'D' in the oplog
we must delete the record, however this means that we cannot generate XML with the details of
the record because it no longer exists.

This commit gets around this issue by treating footnoteassociatemeasures in a special manner during
the generation of the xml and assigning the operation type specifically to be a delete operation if the
operation recorded in the oplog is an update. This is because there is actually no case in which this object
will be updated, any update should be a delete.