Fix: Server errors generated by incomplete form for create footnotes
This commit fixes a bug which caused a server error when a user added an invalid
measure SID for a footnote type that accepted measures, but then switched to one
that didn't without clearing the field.

Before this commit it would raise a server error, this commit allows user to submit the
footnote as expected.