Fix: Withdraw additional code from workbasket causing error
Prior to this change, if user tried to withdraw additional code(s) from a workbasket
in a way that would leave the workbasket empty, errors were being raised. This was
due to the workbasket no longer existing once it's contents had been deleted.

This caused issues in the measures/bulks_controller when some logic depended on
data from the now non existent workbasket.

This commit fixes these issues.