Fix: Cancel edit quota bug
Prior to this fix, when a user selected a quota to edit, but decided to cancel before creating the workbasket
the app crashed. This was due to an incorrect link on the cancel button.

This fix corrects this issue by redirecting the user to the home page after cancelling the edit quota
workbasket.