Fix: Page crashing when viewing edited footnote as cc or reviewer
Prior to this change, when a cross checker or approver approves an edit
footnote workbasket but clicks on the 'view footnote' link the page crashes.

This is due to the incorrect linking of the views. When clicking view footnote,
it was taking the user to the summary of configuration page for a create
footnote workbasket which is incompatible with the page for an edit footnote
workbasket

This commit fixes this issue.