Fix: Change DIT XML file naming convention
The filenaming convention has been agreed by HMRC and European Dynamics,
and the decision has been taken to omit the timestamp, which had been
included in the HMRC interface documentation. Therefore the filenaming
conventions are as follows:

For the Taric 3 XML file:
Using the example of the file from 2019 with an envelope ID of 142, the
filename will be DIT190142.xml, i.e.
- DIT, followed by
- the year (two digits), followed by
- the envelope ID (pre-padded with zeroes up to 4 digits), followed by
- .xml

For the metadata file:
As above, but with _metadata inserted before the .xml extension