New: Upload workbasket XML automatically upon approval
Prior to this change, a user had to manually upload xml to our s3 bucket in order to be processed
by CDS.

This change uploads the xml once the workbasket is approved.

Note: This commit makes the upload run synchronously, i.e not in a sidekiq queue in order for us to
be able to more easily monitor/debug the process. This may need to be change to run async again
once the concept has been proven and any bugs ironed out.