Fix: Withdrawing an edit reglation workbasket bug.
Prior to this change, when withdrawing a workbasket it was creating an
entry in the oplog for deleting the whole regulation.

This change deletes all the entries in oplog connected to this workbasket
rather than adding a delete entry. Also, hides the 'edit the ID' feature
as ID cannot be edited for regulations.

https://uktrade.atlassian.net/browse/TARIFFS-470