Fix: Edit footnote after withdrawing bug
Prior to this change, if a user edited a gootnote, using a description validity start date that already existed,
then withdrew the workbasket and tried to submit again, the user would see a server error.

This was due to the withdraw process deleting the footnote description that was being edited so the second time
the user went to edit, it was not available to update.

This commit fixes this issue by stopping the footnote descriptions being deleted upon withdrawal.