Fix: Additional code approval failing
Prior to this change, approving an additional code sometimes took user to a crash page.

This seems to be because the export worker is running synchronously and if the next page loads
before the job has finished the page crashes.

This change makes the worker run async which should ensure the user never gets to a crash page
for this reason.