Fix: Training crash due to reference bug
Prior to this commit, our training environment was crashing due to use not having a trade tariff reference app in
our training environment that we can link to. The link was looking for an env variable which did not exist which
caused the app to crash, this commit fixes this issue by only showing the link if the ENV variable exists.

This is a short term fix, in the long term we need to set up a reference app in the training environment.