Fix: Edit latest Quota order number
Prior to this change, when a user selects a quota order number to edit they sometimes get an error
saying that the quota order number has an end date which means it is not editable, when it does not
have an end date.

This change ensures that the user is always editing the latest quota order number period.