Fix: Withdraw/edit edit quota suspension
Prior to this change, a user could not withdraw/edit an edit quota suspension workbasket.

This was due to 2 things. Firstly the withdraw link did not work properly, secondly the necessary data was not
persisted in the workbasket which would allow the user to continue where they had left off before withdrawing.

This commit fixes this issue.