Fix: Ensure new nomenclatures match with a Chapter
Prior to this change, it was possible to create a nomenclture that could
not be displayed as it didn't correspond to a vaild section/chapter.

This change makes sure that one exists. Also, this change will display a
nomenclature item that had previously been 'orphaned' due to it's indent
being a higher value than expected. (E.g. If an item had an indent of 3
and the next item had an indent of 5, then the item with 5 would not
have been displayed as it was considered to be detached. This is now
displayed so that it can be deleted if necessary.

https://uktrade.atlassian.net/browse/TARIFFS-639