Fix: Edit/Delete quota suspension period xml bug
Prior to this change, both edit and delete quota suspension workbaskets are rendering the XML for a delete
transaction.

This is due `record_is_a_pending_deletion?` returning true anytime the model in question is a quota
suspension period.

This change instead looks at the type of workbasket and only returns true if the workbasket type is a
DeleteQuotaSuspensionSettings which fixes the issue.