Fix: Create quota blocking period date validation
Prior to this commit, a user could select a date in the past and could not select a blocking period to start and end on the same
day.

Users should only be allowed to start a blocking period in the future, and the accepted way to create a one day blocking period
in the current system is to have the start and end date on the same day so this should be allowed.

This commit fixes these issues.